fix: prevent focusout replay recursion in radix dialog compatibility
