+++ src/widget/index.ts
@@ -29,6 +29,7 @@ import {
   startConsoleLogCapture,
   type ConsoleLogEntry,
 } from './console-logs';
+import { installRadixDialogCompatibility } from './radix-compat';
 import { resolveAccentColor } from '../defaults';
 
 type FeedbackCategory = 'bug' | 'feature' | 'question';
@@ -761,96 +762,6 @@ function createPullTab(root: HTMLElement, config: WidgetConfig): HTMLElement {
   root.appendChild(tab);
   _pullTab = tab;
   return tab;
-}
-
-function installRadixDialogCompatibility(host: HTMLElement): void {
-  const preventBugDropDismissal = (event: Event) => {
-    if (isBugDropInteraction(host, event)) {
-      event.preventDefault();
-    }
-  };
-  const keepBugDropFocus = (event: Event) => {
-    if (isBugDropFocusEvent(host, event)) {
-      if (event.type === 'focusin') {
-        event.stopImmediatePropagation();
-        return;
-      }
-
-      if (event.type === 'focusout') {
-        replayHostFocusOut(event);
-        event.stopImmediatePropagation();
-        return;
-      }
-
-      event.stopImmediatePropagation();
-    }
-  };
-
-  for (const eventType of [
-    'dismissableLayer.pointerDownOutside',
-    'dismissableLayer.interactOutside',
-  ] as const) {
-    document.addEventListener(eventType, preventBugDropDismissal, true);
-  }
-  window.addEventListener('focusin', keepBugDropFocus, true);
-  window.addEventListener('focusout', keepBugDropFocus, true);
-}
-
-function isBugDropInteraction(host: HTMLElement, event: Event): boolean {
-  const originalEvent = (event as CustomEvent<{ originalEvent?: Event }>).detail?.originalEvent;
-  const path =
-    typeof originalEvent?.composedPath === 'function'
-      ? originalEvent.composedPath()
-      : typeof event.composedPath === 'function'
-        ? event.composedPath()
-        : [];
-
-  if (path.includes(host)) {
-    return true;
-  }
-
-  return (originalEvent?.target ?? event.target) === host;
-}
-
-function isBugDropFocusEvent(host: HTMLElement, event: Event): boolean {
-  if (!(event instanceof FocusEvent)) {
-    return isBugDropInteraction(host, event);
-  }
-
-  if (event.type === 'focusin') {
-    return isBugDropInteraction(host, event);
-  }
-
-  if (event.type !== 'focusout') {
-    return false;
-  }
-
-  const nextFocusedNode = event.relatedTarget;
-  const nextFocusIsBugDrop =
-    nextFocusedNode === host ||
-    (nextFocusedNode instanceof Node && (host.shadowRoot?.contains(nextFocusedNode) ?? false));
-  return nextFocusIsBugDrop && !isBugDropInteraction(host, event);
-}
-
-function replayHostFocusOut(event: Event): void {
-  const path = typeof event.composedPath === 'function' ? event.composedPath() : [];
-  for (const node of path) {
-    if (!(node instanceof HTMLElement)) {
-      continue;
-    }
-
-    node.dispatchEvent(
-      new FocusEvent('focusout', {
-        bubbles: false,
-        composed: false,
-        relatedTarget: event instanceof FocusEvent ? event.relatedTarget : null,
-      })
-    );
-
-    if (node === document.body) {
-      break;
-    }
-  }
 }
 
 function initWidget(config: WidgetConfig) {
